Add presentation options for prefered navigation bar hidden state
To be used when the presenting view controller is a navigation controller and we want to hide its navigation bar. We added 2 separate options so that we act on them only if they are explicitly provided to not override any existing customizations made inside presentables.

The options are internal so that they can be used only through the static function which implies that you should have only one of them at a time (but can be workarounded of course).